+++ tests/src/cli/test_cli_actions_class.py
@@ -3,6 +3,6 @@ from cli import CLIActions
 
 class TestChoicesMethod:
     actual_choices = CLIActions.choices()
-    expected_choices = [CLIActions.SEARCH, CLIActions.FETCH, CLIActions.SETTINGS_LIST]
+    expected_choices = [CLIActions.SEARCH, CLIActions.FETCH, CLIActions.SETTINGS_LIST, CLIActions.TOTP]
 
     assert actual_choices == expected_choices